+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [1.0.0] - 2021-04-15
+
+## Changed
+- use `dataTimestamp` as part of consensus
+- use `chain.getStatus` to pull all consensus data
+
+## Removed
+- `ValidatorRegistry`
+
 ## [0.10.6] - 2021-04-15
 
 ### Fix